- moved manipulator enums into manipulationflags
- changed manipulator enums type to actual bitflag
- exposed bitflag enum types as blueprint type so we can use them in bitmask nodes
  - added hidden zero entry to bitmasks - unreal won't allow blueprint types without zero value
- changed datatype of properties using manpulationflags from uint8 to int32 - sticking with uint8 will make the property a bitmask byte type in blueprints and won't allow assigning of multiple values and bitmask nodes whereas int32 will define them as integer bitmask values that allows the usual toggling and assignment of bitmask nodes

Related work items: #468

+++ UXToolsGame/Plugins/UXTools/Source/UXTools/Public/Interactions/UxtGrabTargetComponent.h
@@ -25,7 +25,7 @@ struct UXTOOLS_API FUxtGrabPointerData
 	UUxtNearPointerComponent* NearPointer = nullptr;
 
 	/** The far pointer that is interacting with the component */
-    UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Grab Pointer Data")
+	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Grab Pointer Data")
 	UUxtFarPointerComponent* FarPointer = nullptr;
 
 	/** Last updated grab point transform. (Pointer transform in near pointer case, ray hit transform in far pointer case) */
@@ -240,7 +240,7 @@ public:
 
 	/** Property that indicates if the object is grabbable with far or near interaction or both. */
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Interaction, meta = (Bitmask, BitmaskEnum = EUxtInteractionMode))
-	uint8 InteractionMode;
+	int32 InteractionMode;
 
 private:
